Improve datatable related code to remove long, double

+++ modules/ballerina-core/src/main/java/org/ballerinalang/model/DataTableJSONDataSource.java
@@ -76,10 +76,10 @@ public class DataTableJSONDataSource implements JSONDataSource {
                     objNode.put(name, df.getString(name));
                     break;
                 case INT:
-                    objNode.put(name, df.getLong(name));
+                    objNode.put(name, df.getInt(name));
                     break;
                 case FLOAT:
-                    objNode.put(name, df.getDouble(name));
+                    objNode.put(name, df.getFloat(name));
                     break;
                 case BOOLEAN:
                     objNode.put(name, df.getBoolean(name));
