+++ modules/ballerina-core/src/main/java/org/ballerinalang/model/DataTableOMDataSource.java
@@ -65,10 +65,10 @@ public class DataTableOMDataSource extends AbstractPushOMDataSource {
                     value = dataTable.getString(col.getName());
                     break;
                 case INT:
-                    value = String.valueOf(dataTable.getLong(col.getName()));
+                    value = String.valueOf(dataTable.getInt(col.getName()));
                     break;
                 case FLOAT:
-                    value = String.valueOf(dataTable.getDouble(col.getName()));
+                    value = String.valueOf(dataTable.getFloat(col.getName()));
                     break;
                 case ARRAY:
                     isArray = true;
