+++ modules/ballerina-native/src/main/java/org/ballerinalang/nativeimpl/lang/datatables/GetIntByIndex.java
@@ -56,6 +56,6 @@ public class GetIntByIndex extends AbstractNativeFunction {
     public BValue[] execute(Context ctx) {
         BDataTable dataTable = (BDataTable) getArgument(ctx, 0);
         int index = ((BInteger) getArgument(ctx, 1)).intValue();
-        return getBValues(new BInteger(dataTable.getLong(index)));
+        return getBValues(new BInteger(dataTable.getInt(index)));
     }
 }
